dom/workers/test/test_readableStream_when_closing.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for ReadableStream+fetch when the worker is closing</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
</head>
<body>
  <script>
SimpleTest.waitForExplicitFinish();

info("Starting test");

const WORKER_URL = "chrome://mochitests/content/chrome/dom/workers/test/file_readableStream_when_closing_worker.js";

const w = new Worker(WORKER_URL);
w.onmessage = function(e) {
  ok(true, 'Worker is reading');

  const wdm = Cc["@mozilla.org/dom/workers/workerdebuggermanager;1"].
              getService(Ci.nsIWorkerDebuggerManager);
  wdm.addListener({
    onUnregister (dbg) {
      if (dbg.url == WORKER_URL) {
        ok(true, "Debugger with url " + WORKER_URL + " should be unregistered.");
        wdm.removeListener(this);
        SimpleTest.finish();
      }
    }
  });

  w.terminate();
}
w.postMessage("start");
  </script>
</body>
</html>
